refactor/rename api crate contents, make crate naming consistent (all starting with grin_wallet)
