+++ Cargo.toml
@@ -16,7 +16,7 @@ name = "grin-wallet"
 path = "src/bin/grin-wallet.rs"
 
 [workspace]
-members = ["apiwallet", "config", "libwallet", "refwallet"]
+members = ["api", "config", "libwallet", "refwallet"]
 exclude = ["integration"]
 
 [dependencies]
@@ -29,9 +29,9 @@ prettytable-rs = "0.7"
 log = "0.4"
 linefeed = "0.5"
 
-grin_apiwallet = { path = "./apiwallet", version = "1.1.0" }
-grin_libwallet = { path = "./libwallet", version = "1.1.0" }
-grin_refwallet = { path = "./refwallet", version = "1.1.0" }
+grin_wallet_api = { path = "./api", version = "1.1.0" }
+grin_wallet_libwallet = { path = "./libwallet", version = "1.1.0" }
+grin_wallet_refwallet = { path = "./refwallet", version = "1.1.0" }
 grin_wallet_config = { path = "./config", version = "1.1.0" }
 
 grin_core = { git = "https://github.com/mimblewimble/grin", branch = "milestone/1.1.0" }
