+++ api/src/owner.rs
@@ -41,501 +41,17 @@ use crate::keychain::{Identifier, Keychain};
 use crate::libwallet::internal::{keys, tx, updater};
 use crate::libwallet::slate::Slate;
 use crate::libwallet::types::{
-	AcctPathMapping, BlockFees, CbData, NodeClient, OutputData, OutputLockFn, TxLogEntry,
-	TxLogEntryType, TxWrapper, WalletBackend, WalletInfo,
+	AcctPathMapping, NodeClient, OutputData, OutputLockFn, TxLogEntry,
+	TxWrapper, WalletBackend, WalletInfo,
 };
 use crate::libwallet::{Error, ErrorKind};
 use crate::util;
 use crate::util::secp::{pedersen, ContextFlag, Secp256k1};
-use easy_jsonrpc;
 
 const USER_MESSAGE_MAX_LEN: usize = 256;
 
-/// Public definition used to generate jsonrpc api for APIOwner.
-#[easy_jsonrpc::rpc]
-pub trait OwnerApi {
-	/**
-	Networked version of [APIOwner::accounts](struct.APIOwner.html#method.accounts).
-
-	# Json rpc example
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "accounts",
-		"params": [],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Ok": [
-				{
-					"label": "default",
-					"path": "0200000000000000000000000000000000"
-				}
-			]
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn accounts(&self) -> Result<Vec<AcctPathMapping>, ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::create_account_path](struct.APIOwner.html#method.create_account_path).
-
-	# Json rpc example
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "create_account_path",
-		"params": ["account1"],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Ok": "0200000001000000000000000000000000"
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn create_account_path(&self, label: &String) -> Result<Identifier, ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::set_active_account](struct.APIOwner.html#method.set_active_account).
-
-	# Json rpc example
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "set_active_account",
-		"params": ["default"],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Ok": null
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn set_active_account(&self, label: &String) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::retrieve_outputs](struct.APIOwner.html#method.retrieve_outputs).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "retrieve_outputs",
-		"params": [false, false, null],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn retrieve_outputs(
-		&self,
-		include_spent: bool,
-		refresh_from_node: bool,
-		tx_id: Option<u32>,
-	) -> Result<(bool, Vec<(OutputData, pedersen::Commitment)>), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::retrieve_txs](struct.APIOwner.html#method.retrieve_txs).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "retrieve_txs",
-		"params": [false, null, null],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn retrieve_txs(
-		&self,
-		refresh_from_node: bool,
-		tx_id: Option<u32>,
-		tx_slate_id: Option<Uuid>,
-	) -> Result<(bool, Vec<TxLogEntry>), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::retrieve_summary_info](struct.APIOwner.html#method.retrieve_summary_info).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "retrieve_summary_info",
-		"params": [false, 1],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn retrieve_summary_info(
-		&self,
-		refresh_from_node: bool,
-		minimum_confirmations: u64,
-	) -> Result<(bool, WalletInfo), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::estimate_initiate_tx](struct.APIOwner.html#method.estimate_initiate_tx).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "estimate_initiate_tx",
-		"params": [null, 0, 0, 10, 0, false],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn estimate_initiate_tx(
-		&self,
-		src_acct_name: Option<String>,
-		amount: u64,
-		minimum_confirmations: u64,
-		max_outputs: usize,
-		num_change_outputs: usize,
-		selection_strategy_is_use_all: bool,
-	) -> Result<(/* total */ u64, /* fee */ u64), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::finalize_tx](struct.APIOwner.html#method.finalize_tx).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "finalize_tx",
-		"params": [{
-			"version_info": {
-				"version": 2,
-				"orig_version": 2,
-				"min_compat_version": 0
-			},
-			"amount": 0,
-			"fee": 0,
-			"height": 0,
-			"id": "414bad48-3386-4fa7-8483-72384c886ba3",
-			"lock_height": 0,
-			"num_participants": 2,
-			"participant_data": [],
-			"tx": {
-				"body": {
-					"inputs": [],
-					"kernels": [],
-					"outputs": []
-				},
-				"offset": "0000000000000000000000000000000000000000000000000000000000000000"
-			}
-		}],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn finalize_tx(&self, slate: Slate) -> Result<Slate, ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::cancel_tx](struct.APIOwner.html#method.cancel_tx).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "cancel_tx",
-		"params": [null, null],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn cancel_tx(&self, tx_id: Option<u32>, tx_slate_id: Option<Uuid>) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::get_stored_tx](struct.APIOwner.html#method.get_stored_tx).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "get_stored_tx",
-		"params": [
-			{
-				"amount_credited": 0,
-				"amount_debited": 0,
-				"confirmation_ts": null,
-				"confirmed": false,
-				"creation_ts": "2019-03-05T20:49:59.444095Z",
-				"fee": null,
-				"id": 10,
-				"messages": null,
-				"num_inputs": 0,
-				"num_outputs": 0,
-				"parent_key_id": "0000000000000000000000000000000000",
-				"stored_tx": null,
-				"tx_slate_id": null,
-				"tx_type": "TxReceived"
-			}
-		],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Ok": null
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn get_stored_tx(&self, entry: &TxLogEntry) -> Result<Option<Transaction>, ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::post_tx](struct.APIOwner.html#method.post_tx).
-
-	```no_run
-    # // This test currently fails on travis
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "post_tx",
-		"params": [
-			{
-				"body": {
-					"inputs": [],
-					"kernels": [],
-					"outputs": []
-				},
-				"offset": "0000000000000000000000000000000000000000000000000000000000000000"
-			},
-			false
-		],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"ClientCallback": "Posting transaction to node: Request error: Cannot make request: an error occurred trying to connect: Connection refused (os error 61)"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn post_tx(&self, tx: &Transaction, fluff: bool) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::verify_slate_messages](struct.APIOwner.html#method.verify_slate_messages).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "verify_slate_messages",
-		"params": [{
-			"version_info": {
-				"version": 2,
-				"orig_version": 2,
-				"min_compat_version": 0
-			},
-			"amount": 0,
-			"fee": 0,
-			"height": 0,
-			"id": "414bad48-3386-4fa7-8483-72384c886ba3",
-			"lock_height": 0,
-			"num_participants": 2,
-			"participant_data": [],
-			"tx": {
-				"body": {
-					"inputs": [],
-					"kernels": [],
-					"outputs": []
-				},
-				"offset": "0000000000000000000000000000000000000000000000000000000000000000"
-			}
-		}],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			 "Ok": null
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn verify_slate_messages(&self, slate: &Slate) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::restore](struct.APIOwner.html#method.restore).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "restore",
-		"params": [],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn restore(&self) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::check_repair](struct.APIOwner.html#method.check_repair).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "check_repair",
-		"params": [false],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn check_repair(&self, delete_unconfirmed: bool) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIOwner::node_height](struct.APIOwner.html#method.node_height).
-
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_owner_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "node_height",
-		"params": [],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-			"Err": {
-				"CallbackImpl": "Error opening wallet"
-			}
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn node_height(&self) -> Result<(u64, bool), ErrorKind>;
-}
-
 /// Functions intended for use by the owner (e.g. master seed holder) of the wallet.
-pub struct APIOwner<W: ?Sized, C, K>
+pub struct Owner<W: ?Sized, C, K>
 where
 	W: WalletBackend<C, K>,
 	C: NodeClient,
@@ -548,7 +64,7 @@ where
 	phantom_c: PhantomData<C>,
 }
 
-impl<W: ?Sized, C, K> APIOwner<W, C, K>
+impl<W: ?Sized, C, K> Owner<W, C, K>
 where
 	W: WalletBackend<C, K>,
 	C: NodeClient,
@@ -580,7 +96,7 @@ where
 	/// use util::Mutex;
 	///
 	/// use keychain::ExtKeychain;
-	/// use wallet::libwallet::api::APIOwner;
+	/// use wallet::libwallet::api::Owner;
 	///
 	/// // These contain sample implementations of each part needed for a wallet
 	/// use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend};
@@ -598,13 +114,13 @@ where
 	///			LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	///		));
 	///
-	/// let api_owner = APIOwner::new(wallet.clone());
+	/// let api_owner = Owner::new(wallet.clone());
 	/// // .. perform wallet operations
 	///
 	/// ```
 
 	pub fn new(wallet_in: Arc<Mutex<W>>) -> Self {
-		APIOwner {
+		Owner {
 			wallet: wallet_in,
 			phantom: PhantomData,
 			phantom_c: PhantomData,
@@ -626,7 +142,7 @@ where
 	/// * This method does not need to use the wallet seed or keychain.
 	///
 	/// # Example
-	/// Set up as in [`new`](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [`new`](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet_config as config;
 	/// # extern crate grin_refwallet as wallet;
@@ -635,7 +151,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend};
 	/// # use config::WalletConfig;
 	/// # let mut wallet_config = WalletConfig::default();
@@ -646,7 +162,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let api_owner = APIOwner::new(wallet.clone());
+	/// let api_owner = Owner::new(wallet.clone());
 	///
 	/// let result = api_owner.accounts();
 	///
@@ -678,13 +194,13 @@ where
 	/// so the first call will create an account at `m/1/0` and the second at
 	/// `m/2/0` etc. . .
 	/// * The account path is used throughout as the parent key for most key-derivation
-	/// operations. See [`set_active_account`](struct.APIOwner.html#method.set_active_account) for
+	/// operations. See [`set_active_account`](struct.Owner.html#method.set_active_account) for
 	/// further details.
 	///
 	/// * This function does not need to use the root wallet seed or keychain.
 	///
 	/// # Example
-	/// Set up as in [`new`](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [`new`](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet as wallet;
 	/// # extern crate grin_keychain as keychain;
@@ -692,7 +208,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend,  WalletConfig};
 	/// # let mut wallet_config = WalletConfig::default();
 	/// # wallet_config.data_file_dir = "test_output/doc/wallet1".to_owned();
@@ -702,7 +218,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let api_owner = APIOwner::new(wallet.clone());
+	/// let api_owner = Owner::new(wallet.clone());
 	///
 	/// let result = api_owner.create_account_path("account1");
 	///
@@ -721,7 +237,7 @@ where
 	///
 	/// # Arguments
 	/// * `label` - The human readable label for the account. Accounts can be retrieved via
-	/// the [`account`](struct.APIOwner.html#method.accounts) method
+	/// the [`account`](struct.Owner.html#method.accounts) method
 	///
 	/// # Returns
 	/// * Result Containing:
@@ -738,7 +254,7 @@ where
 	/// * This function does not need to use the root wallet seed or keychain.
 	///
 	/// # Example
-	/// Set up as in [`new`](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [`new`](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet as wallet;
 	/// # extern crate grin_keychain as keychain;
@@ -746,7 +262,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend,  WalletConfig};
 	/// # let mut wallet_config = WalletConfig::default();
 	/// # wallet_config.data_file_dir = "test_output/doc/wallet1".to_owned();
@@ -756,7 +272,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let api_owner = APIOwner::new(wallet.clone());
+	/// let api_owner = Owner::new(wallet.clone());
 	///
 	/// let result = api_owner.create_account_path("account1");
 	///
@@ -796,7 +312,7 @@ where
 	/// and the Output commitment as identified in the chain's UTXO set
 	///
 	/// # Example
-	/// Set up as in [`new`](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [`new`](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet as wallet;
 	/// # extern crate grin_keychain as keychain;
@@ -804,7 +320,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend,  WalletConfig};
 	/// # let mut wallet_config = WalletConfig::default();
 	/// # wallet_config.data_file_dir = "test_output/doc/wallet1".to_owned();
@@ -814,7 +330,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let api_owner = APIOwner::new(wallet.clone());
+	/// let api_owner = Owner::new(wallet.clone());
 	/// let show_spent = false;
 	/// let update_from_node = true;
 	/// let tx_id = None;
@@ -873,7 +389,7 @@ where
 	/// [TxLogEntries](../types/struct/TxLogEntry.html)
 	///
 	/// # Example
-	/// Set up as in [`new`](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [`new`](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet as wallet;
 	/// # extern crate grin_keychain as keychain;
@@ -881,7 +397,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend,  WalletConfig};
 	/// # let mut wallet_config = WalletConfig::default();
 	/// # wallet_config.data_file_dir = "test_output/doc/wallet1".to_owned();
@@ -891,7 +407,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let api_owner = APIOwner::new(wallet.clone());
+	/// let api_owner = Owner::new(wallet.clone());
 	/// let update_from_node = true;
 	/// let tx_id = None;
 	/// let tx_slate_id = None;
@@ -947,7 +463,7 @@ where
 	/// * The second element contains the Summary [`WalletInfo`](../types/struct.WalletInfo.html)
 	///
 	/// # Example
-	/// Set up as in [`new`](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [`new`](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet as wallet;
 	/// # extern crate grin_keychain as keychain;
@@ -955,7 +471,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend,  WalletConfig};
 	/// # let mut wallet_config = WalletConfig::default();
 	/// # wallet_config.data_file_dir = "test_output/doc/wallet1".to_owned();
@@ -965,7 +481,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let mut api_owner = APIOwner::new(wallet.clone());
+	/// let mut api_owner = Owner::new(wallet.clone());
 	/// let update_from_node = true;
 	/// let minimum_confirmations=10;
 	///
@@ -1009,7 +525,7 @@ where
 	/// attempt to re-spend outputs that are already included in a transaction before the transaction
 	/// is confirmed. This method also returns a function that will perform that locking, and it is
 	/// up to the caller to decide the best time to call the lock function
-	/// (via the [`tx_lock_outputs`](struct.APIOwner.html#method.tx_lock_outputs) method).
+	/// (via the [`tx_lock_outputs`](struct.Owner.html#method.tx_lock_outputs) method).
 	/// If the exchange method is intended to be synchronous (such as via a direct http call,)
 	/// then the lock call can wait until the response is confirmed. If it is asynchronous, (such
 	/// as via file transfer,) the lock call should happen immediately (before the file is sent
@@ -1018,7 +534,7 @@ where
 	/// # Arguments
 	/// * `src_acct_name` - The human readable account name from which to draw outputs
 	/// for the transaction, overriding whatever the active account is as set via the
-	/// [`set_active_account`](struct.APIOwner.html#method.set_active_account) method.
+	/// [`set_active_account`](struct.Owner.html#method.set_active_account) method.
 	/// If None, the transaction will use the active account.
 	/// * `amount` - The amount to send, in nanogrins. (`1 G = 1_000_000_000nG`)
 	/// * `minimum_confirmations` - The minimum number of confirmations an output
@@ -1051,7 +567,7 @@ where
 	/// * A lock function, which should be called when the caller deems it appropriate to lock
 	/// the transaction outputs (i.e. there is relative certaintly that the slate will be
 	/// transmitted to the receiving party). Must be called before calling
-	/// [`finalize_tx`](struct.APIOwner.html#method.finalize_tx).
+	/// [`finalize_tx`](struct.Owner.html#method.finalize_tx).
 	/// * or [`libwallet::Error`](../struct.Error.html) if an error is encountered.
 	///
 	/// # Remarks
@@ -1059,10 +575,10 @@ where
 	/// * This method requires an active connection to a node, and will fail with error if a node
 	/// cannot be contacted to refresh output statuses.
 	/// * This method will store a partially completed transaction in the wallet's transaction log,
-	/// which will be updated on the corresponding call to [`finalize_tx`](struct.APIOwner.html#method.finalize_tx).
+	/// which will be updated on the corresponding call to [`finalize_tx`](struct.Owner.html#method.finalize_tx).
 	///
 	/// # Example
-	/// Set up as in [new](struct.APIOwner.html#method.new) method above.
+	/// Set up as in [new](struct.Owner.html#method.new) method above.
 	/// ``` ignore
 	/// # extern crate grin_wallet as wallet;
 	/// # extern crate grin_keychain as keychain;
@@ -1070,7 +586,7 @@ where
 	/// # use std::sync::Arc;
 	/// # use util::Mutex;
 	/// # use keychain::ExtKeychain;
-	/// # use wallet::libwallet::api::APIOwner;
+	/// # use wallet::libwallet::api::Owner;
 	/// # use wallet::{LMDBBackend, HTTPNodeClient, WalletBackend,  WalletConfig};
 	/// # let mut wallet_config = WalletConfig::default();
 	/// # wallet_config.data_file_dir = "test_output/doc/wallet1".to_owned();
@@ -1080,7 +596,7 @@ where
 	/// # 	LMDBBackend::new(wallet_config.clone(), "", node_client).unwrap()
 	/// # ));
 	///
-	/// let mut api_owner = APIOwner::new(wallet.clone());
+	/// let mut api_owner = Owner::new(wallet.clone());
 	/// let amount = 2_000_000_000;
 	///
 	/// // Attempt to create a transaction using the 'default' account
@@ -1169,7 +685,7 @@ where
 	/// # Arguments
 	/// * `src_acct_name` - The human readable account name from which to draw outputs
 	/// for the transaction, overriding whatever the active account is as set via the
-	/// [`set_active_account`](struct.APIOwner.html#method.set_active_account) method.
+	/// [`set_active_account`](struct.Owner.html#method.set_active_account) method.
 	/// If None, the transaction will use the active account.
 	/// * `amount` - The amount to send, in nanogrins. (`1 G = 1_000_000_000nG`)
 	/// * `minimum_confirmations` - The minimum number of confirmations an output
@@ -1364,480 +880,4 @@ where
 			Err(_) => false,
 		}
 	}
-}
-
-impl<W: ?Sized, C, K> OwnerApi for APIOwner<W, C, K>
-where
-	W: WalletBackend<C, K>,
-	C: NodeClient,
-	K: Keychain,
-{
-	fn accounts(&self) -> Result<Vec<AcctPathMapping>, ErrorKind> {
-		APIOwner::accounts(self).map_err(|e| e.kind())
-	}
-
-	fn create_account_path(&self, label: &String) -> Result<Identifier, ErrorKind> {
-		APIOwner::create_account_path(self, label).map_err(|e| e.kind())
-	}
-
-	fn set_active_account(&self, label: &String) -> Result<(), ErrorKind> {
-		APIOwner::set_active_account(self, label).map_err(|e| e.kind())
-	}
-
-	fn retrieve_outputs(
-		&self,
-		include_spent: bool,
-		refresh_from_node: bool,
-		tx_id: Option<u32>,
-	) -> Result<(bool, Vec<(OutputData, pedersen::Commitment)>), ErrorKind> {
-		APIOwner::retrieve_outputs(self, include_spent, refresh_from_node, tx_id)
-			.map_err(|e| e.kind())
-	}
-
-	fn retrieve_txs(
-		&self,
-		refresh_from_node: bool,
-		tx_id: Option<u32>,
-		tx_slate_id: Option<Uuid>,
-	) -> Result<(bool, Vec<TxLogEntry>), ErrorKind> {
-		APIOwner::retrieve_txs(self, refresh_from_node, tx_id, tx_slate_id).map_err(|e| e.kind())
-	}
-
-	fn retrieve_summary_info(
-		&self,
-		refresh_from_node: bool,
-		minimum_confirmations: u64,
-	) -> Result<(bool, WalletInfo), ErrorKind> {
-		APIOwner::retrieve_summary_info(self, refresh_from_node, minimum_confirmations)
-			.map_err(|e| e.kind())
-	}
-
-	fn estimate_initiate_tx(
-		&self,
-		src_acct_name: Option<String>,
-		amount: u64,
-		minimum_confirmations: u64,
-		max_outputs: usize,
-		num_change_outputs: usize,
-		selection_strategy_is_use_all: bool,
-	) -> Result<(/* total */ u64, /* fee */ u64), ErrorKind> {
-		APIOwner::estimate_initiate_tx(
-			self,
-			src_acct_name.as_ref().map(String::as_str),
-			amount,
-			minimum_confirmations,
-			max_outputs,
-			num_change_outputs,
-			selection_strategy_is_use_all,
-		)
-		.map_err(|e| e.kind())
-	}
-
-	fn finalize_tx(&self, mut slate: Slate) -> Result<Slate, ErrorKind> {
-		APIOwner::finalize_tx(self, &mut slate).map_err(|e| e.kind())?;
-		Ok(slate)
-	}
-
-	fn cancel_tx(&self, tx_id: Option<u32>, tx_slate_id: Option<Uuid>) -> Result<(), ErrorKind> {
-		APIOwner::cancel_tx(self, tx_id, tx_slate_id).map_err(|e| e.kind())
-	}
-
-	fn get_stored_tx(&self, entry: &TxLogEntry) -> Result<Option<Transaction>, ErrorKind> {
-		APIOwner::get_stored_tx(self, entry).map_err(|e| e.kind())
-	}
-
-	fn post_tx(&self, tx: &Transaction, fluff: bool) -> Result<(), ErrorKind> {
-		APIOwner::post_tx(self, tx, fluff).map_err(|e| e.kind())
-	}
-
-	fn verify_slate_messages(&self, slate: &Slate) -> Result<(), ErrorKind> {
-		APIOwner::verify_slate_messages(self, slate).map_err(|e| e.kind())
-	}
-
-	fn restore(&self) -> Result<(), ErrorKind> {
-		APIOwner::restore(self).map_err(|e| e.kind())
-	}
-
-	fn check_repair(&self, delete_unconfirmed: bool) -> Result<(), ErrorKind> {
-		APIOwner::check_repair(self, delete_unconfirmed).map_err(|e| e.kind())
-	}
-
-	fn node_height(&self) -> Result<(u64, bool), ErrorKind> {
-		APIOwner::node_height(self).map_err(|e| e.kind())
-	}
-}
-
-/// Public definition used to generate jsonrpc api for APIForeign.
-#[easy_jsonrpc::rpc]
-pub trait ForeignApi {
-	/**
-	Networked version of [APIForeign::build_coinbase](struct.APIForeign.html#method.build_coinbase).
-
-	# Json rpc example
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_foreign_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "build_coinbase",
-		"params": [
-            {
-                "fees": 0,
-            	"height": 0,
-                "key_id": null
-            }
-        ],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-            "Err": {
-                "CallbackImpl": "Error opening wallet"
-            }
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn build_coinbase(&self, block_fees: &BlockFees) -> Result<CbData, ErrorKind>;
-
-	/**
-	Networked version of [APIForeign::verify_slate_messages](struct.APIForeign.html#method.verify_slate_messages).
-
-	# Json rpc example
-
-	```
-	# grin_apiwallet::doctest_helper_json_rpc_foreign_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "verify_slate_messages",
-		"params": [
-            {
-    			"version_info": {
-    				"version": 2,
-    				"orig_version": 2,
-    				"min_compat_version": 0
-    			},
-    			"amount": 0,
-    			"fee": 0,
-    			"height": 0,
-    			"id": "414bad48-3386-4fa7-8483-72384c886ba3",
-    			"lock_height": 0,
-    			"num_participants": 2,
-    			"participant_data": [],
-    			"tx": {
-    				"body": {
-    					"inputs": [],
-    					"kernels": [],
-    					"outputs": []
-    				},
-    				"offset": "0000000000000000000000000000000000000000000000000000000000000000"
-    			}
-		    }
-        ],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-            "Ok": null
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn verify_slate_messages(&self, slate: &Slate) -> Result<(), ErrorKind>;
-
-	/**
-	Networked version of [APIForeign::receive_tx](struct.APIForeign.html#method.receive_tx).
-
-	# Json rpc example
-
-	```ignore //TODO: No idea why this isn't expanding properly, check as we adjust the API
-	# grin_apiwallet::doctest_helper_json_rpc_foreign_assert_response!(
-	{
-		"jsonrpc": "2.0",
-		"method": "receive_tx",
-		"params": [
-            {
-    			"version_info": {
-    				"version": 2,
-    				"orig_version": 2,
-    				"min_compat_version": 0
-    			},
-    			"amount": 0,
-    			"fee": 0,
-    			"height": 0,
-    			"id": "414bad48-3386-4fa7-8483-72384c886ba3",
-    			"lock_height": 0,
-    			"num_participants": 2,
-    			"participant_data": [],
-    			"tx": {
-    				"body": {
-    					"inputs": [],
-    					"kernels": [],
-    					"outputs": []
-    				},
-    				"offset": "0000000000000000000000000000000000000000000000000000000000000000"
-    			}
-		    },
-            null,
-            null
-        ],
-		"id": 1
-	},
-	{
-		"jsonrpc": "2.0",
-		"result": {
-            "Err": {
-                "CallbackImpl": "Error opening wallet"
-            }
-		},
-		"id": 1
-	}
-	# );
-	```
-	 */
-	fn receive_tx(
-		&self,
-		slate: Slate,
-		dest_acct_name: Option<String>,
-		message: Option<String>,
-	) -> Result<Slate, ErrorKind>;
-}
-
-/// Wrapper around external API functions, intended to communicate
-/// with other parties
-pub struct APIForeign<W: ?Sized, C, K>
-where
-	W: WalletBackend<C, K>,
-	C: NodeClient,
-	K: Keychain,
-{
-	/// Wallet, contains its keychain (TODO: Split these up into 2 traits
-	/// perhaps)
-	pub wallet: Arc<Mutex<W>>,
-	phantom: PhantomData<K>,
-	phantom_c: PhantomData<C>,
-}
-
-impl<'a, W: ?Sized, C, K> APIForeign<W, C, K>
-where
-	W: WalletBackend<C, K>,
-	C: NodeClient,
-	K: Keychain,
-{
-	/// Create new API instance
-	pub fn new(wallet_in: Arc<Mutex<W>>) -> Box<Self> {
-		Box::new(APIForeign {
-			wallet: wallet_in,
-			phantom: PhantomData,
-			phantom_c: PhantomData,
-		})
-	}
-
-	/// Build a new (potential) coinbase transaction in the wallet
-	pub fn build_coinbase(&self, block_fees: &BlockFees) -> Result<CbData, Error> {
-		let mut w = self.wallet.lock();
-		w.open_with_credentials()?;
-		let res = updater::build_coinbase(&mut *w, block_fees);
-		w.close()?;
-		res
-	}
-
-	/// Verifies all messages in the slate match their public keys
-	pub fn verify_slate_messages(&self, slate: &Slate) -> Result<(), Error> {
-		let secp = Secp256k1::with_caps(ContextFlag::VerifyOnly);
-		slate.verify_messages(&secp)?;
-		Ok(())
-	}
-
-	/// Receive a transaction from a sender
-	pub fn receive_tx(
-		&self,
-		slate: &mut Slate,
-		dest_acct_name: Option<&str>,
-		message: Option<String>,
-	) -> Result<(), Error> {
-		let mut w = self.wallet.lock();
-		w.open_with_credentials()?;
-		let parent_key_id = match dest_acct_name {
-			Some(d) => {
-				let pm = w.get_acct_path(d.to_owned())?;
-				match pm {
-					Some(p) => p.path,
-					None => w.parent_key_id(),
-				}
-			}
-			None => w.parent_key_id(),
-		};
-		// Don't do this multiple times
-		let tx = updater::retrieve_txs(&mut *w, None, Some(slate.id), Some(&parent_key_id), false)?;
-		for t in &tx {
-			if t.tx_type == TxLogEntryType::TxReceived {
-				return Err(ErrorKind::TransactionAlreadyReceived(slate.id.to_string()).into());
-			}
-		}
-
-		let message = match message {
-			Some(mut m) => {
-				m.truncate(USER_MESSAGE_MAX_LEN);
-				Some(m)
-			}
-			None => None,
-		};
-
-		let (_, mut create_fn) =
-			tx::add_output_to_slate(&mut *w, slate, &parent_key_id, 1, message)?;
-		create_fn(&mut *w, &slate.tx, PhantomData, PhantomData)?;
-		tx::update_message(&mut *w, slate)?;
-		w.close()?;
-		Ok(())
-	}
-}
-
-impl<W: ?Sized, C, K> ForeignApi for APIForeign<W, C, K>
-where
-	W: WalletBackend<C, K>,
-	C: NodeClient,
-	K: Keychain,
-{
-	fn build_coinbase(&self, block_fees: &BlockFees) -> Result<CbData, ErrorKind> {
-		APIForeign::build_coinbase(self, block_fees).map_err(|e| e.kind())
-	}
-
-	fn verify_slate_messages(&self, slate: &Slate) -> Result<(), ErrorKind> {
-		APIForeign::verify_slate_messages(self, slate).map_err(|e| e.kind())
-	}
-
-	fn receive_tx(
-		&self,
-		mut slate: Slate,
-		dest_acct_name: Option<String>,
-		message: Option<String>,
-	) -> Result<Slate, ErrorKind> {
-		APIForeign::receive_tx(
-			self,
-			&mut slate,
-			dest_acct_name.as_ref().map(String::as_str),
-			message,
-		)
-		.map_err(|e| e.kind())?;
-		Ok(slate)
-	}
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! doctest_helper_json_rpc_owner_assert_response {
-	($request:tt, $expected_response:tt) => {
-		// create temporary wallet, run jsonrpc request on owner api of wallet, delete wallet, return
-		// json response.
-		// In order to prevent leaking tempdirs, This function should not panic.
-		fn rpc_owner_result(
-			request: serde_json::Value,
-		) -> Result<Option<serde_json::Value>, String> {
-			use easy_jsonrpc::Handler;
-			use grin_apiwallet::api::{APIOwner, OwnerApi};
-			use grin_keychain::ExtKeychain;
-			use grin_refwallet::{HTTPNodeClient, LMDBBackend, WalletBackend};
-			use grin_util::Mutex;
-			use grin_wallet_config::WalletConfig;
-			use serde_json;
-			use std::sync::Arc;
-			use tempfile::tempdir;
-
-			let dir = tempdir().map_err(|e| format!("{:#?}", e))?;
-				{
-				let mut wallet_config = WalletConfig::default();
-				wallet_config.data_file_dir = dir
-					.path()
-					.to_str()
-					.ok_or("Failed to convert tmpdir path to string.".to_owned())?
-					.to_owned();
-				let node_client =
-					HTTPNodeClient::new(&wallet_config.check_node_api_http_addr, None);
-				let wallet: Arc<Mutex<WalletBackend<HTTPNodeClient, ExtKeychain>>> =
-					Arc::new(Mutex::new(
-						LMDBBackend::new(wallet_config.clone(), "", node_client)
-							.map_err(|e| format!("{:#?}", e))?,
-					));
-				let api_owner = APIOwner::new(wallet);
-				let owner_api = &api_owner as &dyn OwnerApi;
-				Ok(owner_api.handle_request(request))
-				}
-			}
-
-		let response = rpc_owner_result(serde_json::json!($request))
-			.unwrap()
-			.unwrap();
-		let expected_response = serde_json::json!($expected_response);
-
-		if response != expected_response {
-			panic!(
-				"(left != right) \nleft: {}\nright: {}",
-				serde_json::to_string_pretty(&response).unwrap(),
-				serde_json::to_string_pretty(&expected_response).unwrap()
-				);
-			}
-	};
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! doctest_helper_json_rpc_foreign_assert_response {
-	($request:tt, $expected_response:tt) => {
-		// create temporary wallet, run jsonrpc request on api of wallet, delete wallet, return
-		// json response.
-		// In order to prevent leaking tempdirs, This function should not panic.
-		fn rpc_owner_result(
-			request: serde_json::Value,
-		) -> Result<Option<serde_json::Value>, String> {
-			use easy_jsonrpc::Handler;
-			use grin_apiwallet::api::{APIForeign, ForeignApi};
-			use grin_keychain::ExtKeychain;
-			use grin_refwallet::{HTTPNodeClient, LMDBBackend, WalletBackend};
-			use grin_util::Mutex;
-			use grin_wallet_config::WalletConfig;
-			use serde_json;
-			use std::sync::Arc;
-			use tempfile::tempdir;
-
-			let dir = tempdir().map_err(|e| format!("{:#?}", e))?;
-				{
-				let mut wallet_config = WalletConfig::default();
-				wallet_config.data_file_dir = dir
-					.path()
-					.to_str()
-					.ok_or("Failed to convert tmpdir path to string.".to_owned())?
-					.to_owned();
-				let node_client =
-					HTTPNodeClient::new(&wallet_config.check_node_api_http_addr, None);
-				let wallet: Arc<Mutex<WalletBackend<HTTPNodeClient, ExtKeychain>>> =
-					Arc::new(Mutex::new(
-						LMDBBackend::new(wallet_config.clone(), "", node_client)
-							.map_err(|e| format!("{:#?}", e))?,
-					));
-				let api_foreign = *APIForeign::new(wallet);
-				let foreign_api = &api_foreign as &dyn ForeignApi;
-				Ok(foreign_api.handle_request(request))
-				}
-			}
-
-		let response = rpc_owner_result(serde_json::json!($request))
-			.unwrap()
-			.unwrap();
-		let expected_response = serde_json::json!($expected_response);
-
-		if response != expected_response {
-			panic!(
-				"(left != right) \nleft: {}\nright: {}",
-				serde_json::to_string_pretty(&response).unwrap(),
-				serde_json::to_string_pretty(&expected_response).unwrap()
-				);
-			}
-	};
 }
