+++ libwallet/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "grin_libwallet"
+name = "grin_wallet_libwallet"
 version = "1.1.0"
 authors = ["Grin Developers <mimblewimble@lists.launchpad.net>"]
 description = "Simple, private and scalable cryptocurrency implementation based on the MimbleWimble chain format."
