+++ refwallet/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "grin_refwallet"
+name = "grin_wallet_refwallet"
 version = "1.1.0"
 authors = ["Grin Developers <mimblewimble@lists.launchpad.net>"]
 description = "Simple, private and scalable cryptocurrency implementation based on the MimbleWimble chain format."
@@ -32,8 +32,8 @@ uuid = { version = "0.6", features = ["serde", "v4"] }
 url = "1.7.0"
 chrono = { version = "0.4.4", features = ["serde"] }
 
-grin_apiwallet = { path = "../apiwallet", version = "1.1.0" }
-grin_libwallet = { path = "../libwallet", version = "1.1.0" }
+grin_wallet_api = { path = "../api", version = "1.1.0" }
+grin_wallet_libwallet = { path = "../libwallet", version = "1.1.0" }
 grin_wallet_config = { path = "../config", version = "1.1.0" }
 
 grin_core = { git = "https://github.com/mimblewimble/grin", branch = "milestone/1.1.0" }
