+++ refwallet/src/lib.rs
@@ -27,8 +27,8 @@ use failure;
 use grin_api as api;
 extern crate grin_core as core;
 use grin_keychain as keychain;
-use grin_libwallet as libwallet;
-use grin_apiwallet as apiwallet;
+use grin_wallet_libwallet as libwallet;
+use grin_wallet_api as apiwallet;
 use grin_store as store;
 use grin_util as util;
 extern crate grin_wallet_config as config;
