+++ refwallet/tests/self_send.rs
@@ -14,7 +14,7 @@
 //! Test a wallet sending to self
 #[macro_use]
 extern crate log;
-extern crate grin_refwallet as wallet;
+extern crate grin_wallet_refwallet as wallet;
 
 use self::core::global;
 use self::core::global::ChainTypes;
@@ -23,7 +23,7 @@ use self::wallet::test_framework::{self, LocalWalletClient, WalletProxy};
 use grin_core as core;
 use grin_keychain as keychain;
 use grin_util as util;
-use grin_libwallet as libwallet;
+use grin_wallet_libwallet as libwallet;
 use std::fs;
 use std::thread;
 use std::time::Duration;
