+++ src/bin/cmd/wallet_args.rs
@@ -20,8 +20,8 @@ use clap::ArgMatches;
 use failure::Fail;
 use grin_core as core;
 use grin_keychain as keychain;
-use grin_refwallet::{command, instantiate_wallet, NodeClient, WalletInst, WalletSeed};
-use grin_refwallet::{Error, ErrorKind};
+use grin_wallet_refwallet::{command, instantiate_wallet, NodeClient, WalletInst, WalletSeed};
+use grin_wallet_refwallet::{Error, ErrorKind};
 use grin_wallet_config::WalletConfig;
 use linefeed::terminal::Signal;
 use linefeed::{Interface, ReadResult};
