+++ src/bin/cmd/wallet_tests.rs
@@ -15,10 +15,10 @@
 #[cfg(test)]
 mod wallet_tests {
 	use clap;
-	use grin_refwallet;
+	use grin_wallet_refwallet;
 	use grin_util as util;
 
-	use grin_refwallet::test_framework::{self, LocalWalletClient, WalletProxy};
+	use grin_wallet_refwallet::test_framework::{self, LocalWalletClient, WalletProxy};
 
 	use clap::{App, ArgMatches};
 	use grin_util::Mutex;
@@ -30,7 +30,7 @@ mod wallet_tests {
 	use grin_core::global;
 	use grin_core::global::ChainTypes;
 	use grin_keychain::ExtKeychain;
-	use grin_refwallet::{LMDBBackend, WalletBackend, WalletInst, WalletSeed};
+	use grin_wallet_refwallet::{LMDBBackend, WalletBackend, WalletInst, WalletSeed};
 	use grin_wallet_config::{GlobalWalletConfig, WalletConfig};
 
 	use super::super::wallet_args;
@@ -49,7 +49,7 @@ mod wallet_tests {
 	pub fn config_command_wallet(
 		dir_name: &str,
 		wallet_name: &str,
-	) -> Result<(), grin_refwallet::Error> {
+	) -> Result<(), grin_wallet_refwallet::Error> {
 		let mut current_dir;
 		let mut default_config = GlobalWalletConfig::default();
 		current_dir = env::current_dir().unwrap_or_else(|e| {
@@ -61,7 +61,7 @@ mod wallet_tests {
 		let mut config_file_name = current_dir.clone();
 		config_file_name.push("grin-wallet.toml");
 		if config_file_name.exists() {
-			return Err(grin_refwallet::ErrorKind::ArgumentError(
+			return Err(grin_wallet_refwallet::ErrorKind::ArgumentError(
 				"grin-wallet.toml already exists in the target directory. Please remove it first"
 					.to_owned(),
 			))?;
@@ -122,7 +122,7 @@ mod wallet_tests {
 		node_client: LocalWalletClient,
 		passphrase: &str,
 		account: &str,
-	) -> Result<Arc<Mutex<WalletInst<LocalWalletClient, ExtKeychain>>>, grin_refwallet::Error> {
+	) -> Result<Arc<Mutex<WalletInst<LocalWalletClient, ExtKeychain>>>, grin_wallet_refwallet::Error> {
 		wallet_config.chain_type = None;
 		// First test decryption, so we can abort early if we have the wrong password
 		let _ = WalletSeed::from_file(&wallet_config, passphrase)?;
@@ -138,7 +138,7 @@ mod wallet_tests {
 		wallet_name: &str,
 		client: &LocalWalletClient,
 		arg_vec: Vec<&str>,
-	) -> Result<String, grin_refwallet::Error> {
+	) -> Result<String, grin_wallet_refwallet::Error> {
 		let args = app.clone().get_matches_from(arg_vec);
 		let _ = get_wallet_subcommand(test_dir, wallet_name, args.clone());
 		let mut config = initial_setup_wallet(test_dir, wallet_name);
@@ -148,7 +148,7 @@ mod wallet_tests {
 	}
 
 	/// command line tests
-	fn command_line_test_impl(test_dir: &str) -> Result<(), grin_refwallet::Error> {
+	fn command_line_test_impl(test_dir: &str) -> Result<(), grin_wallet_refwallet::Error> {
 		setup(test_dir);
 		// Create a new proxy to simulate server and wallet responses
 		let mut wallet_proxy: WalletProxy<LocalWalletClient, ExtKeychain> =
@@ -240,7 +240,7 @@ mod wallet_tests {
 		// Mine a bit into wallet 1 so we have something to send
 		// (TODO: Be able to stop listeners so we can test this better)
 		let wallet1 = instantiate_wallet(config1.clone(), client1.clone(), "password", "default")?;
-		grin_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
+		grin_wallet_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
 			api.set_active_account("mining")?;
 			Ok(())
 		})?;
@@ -313,7 +313,7 @@ mod wallet_tests {
 		let wallet1 = instantiate_wallet(config1.clone(), client1.clone(), "password", "default")?;
 
 		// Check our transaction log, should have 10 entries
-		grin_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
+		grin_wallet_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
 			api.set_active_account("mining")?;
 			let (refreshed, txs) = api.retrieve_txs(true, None, None)?;
 			assert!(refreshed);
@@ -340,7 +340,7 @@ mod wallet_tests {
 
 		// check results in wallet 2
 		let wallet2 = instantiate_wallet(config2.clone(), client2.clone(), "password", "default")?;
-		grin_refwallet::controller::owner_single_use(wallet2.clone(), |api| {
+		grin_wallet_refwallet::controller::owner_single_use(wallet2.clone(), |api| {
 			api.set_active_account("account_1")?;
 			let (_, wallet1_info) = api.retrieve_summary_info(true, 1)?;
 			assert_eq!(wallet1_info.last_confirmed_height, bh);
@@ -396,7 +396,7 @@ mod wallet_tests {
 		// Check our transaction log, should have bh entries + one for the self receive
 		let wallet1 = instantiate_wallet(config1.clone(), client1.clone(), "password", "default")?;
 
-		grin_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
+		grin_wallet_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
 			api.set_active_account("mining")?;
 			let (refreshed, txs) = api.retrieve_txs(true, None, None)?;
 			assert!(refreshed);
@@ -430,7 +430,7 @@ mod wallet_tests {
 		// Check our transaction log, should have bh entries + 2 for the self receives
 		let wallet1 = instantiate_wallet(config1.clone(), client1.clone(), "password", "default")?;
 
-		grin_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
+		grin_wallet_refwallet::controller::owner_single_use(wallet1.clone(), |api| {
 			api.set_active_account("mining")?;
 			let (refreshed, txs) = api.retrieve_txs(true, None, None)?;
 			assert!(refreshed);
